+++ src/app/layout.tsx
@@ -2,7 +2,7 @@ import type { Metadata } from 'next';
 import { Inter } from 'next/font/google';
 import './globals.css';
 import { ReactQueryClientProvider } from '@/providers/query.provider';
-import ClientLayout from './ClientLayout';
+import Header from '@/components/molecules/Header';
 
 const inter = Inter({ subsets: ['latin'] });
 
@@ -19,9 +19,7 @@ export default function RootLayout({
   return (
     <ReactQueryClientProvider>
       <html lang="en">
-        <body className={inter.className}>
-          <ClientLayout>{children}</ClientLayout>
-        </body>
+        <body className={inter.className}>{children}</body>
       </html>
     </ReactQueryClientProvider>
   );
